refactor: subfolders for parts of language server
