+++ DSL-langium/langium-config.json
@@ -2,7 +2,7 @@
     "projectName": "SafeDs",
     "languages": [{
         "id": "safe-ds",
-        "grammar": "src/language-server/safe-ds.langium",
+        "grammar": "src/language-server/grammar/safe-ds.langium",
         "fileExtensions": [".sdspipe", ".sdsschema", ".sdsstub", ".sdstest"],
         "textMate": {
             "out": "syntaxes/safe-ds.tmLanguage.json"
