+++ DSL-langium/src/language-server/grammar/safe-ds.langium
